♻️ Refactor: separate Dateprovider
